fix compile error for GCC due to resolution of uint64_t type.

+++ src/neurons/flif_diffusive.h
@@ -41,6 +41,7 @@ extern "C" {
 #endif
 
 #include <stddef.h>
+#include <stdint.h>
 
 struct flif_diffusive_neuron {
     // Parameters (double precision)
@@ -64,7 +65,7 @@ struct flif_diffusive_neuron {
     double g_euler;           // dt / tau_m
 
     // RNG state (SplitMix64)
-    unsigned long long rng;
+    uint64_t rng;
 };
 
 // API (double throughout)
@@ -73,7 +74,7 @@ void   update_flif_diffusive(struct flif_diffusive_neuron* n, double input, doub
 void   free_flif_diffusive(struct flif_diffusive_neuron* n);
 
 // Optional deterministic seeding for reproducibility
-void   flif_diffusive_set_seed(struct flif_diffusive_neuron* n, unsigned long long seed);
+void   flif_diffusive_set_seed(struct flif_diffusive_neuron* n, uint64_t seed);
 
 #ifdef __cplusplus
 }
